pyproject.toml
[project]
name = "mlops-hotel-reservation"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "google-cloud-storage>=3.2.0",
    "numpy>=2.3.2",
    "pandas>=2.3.1",
    "pyyaml>=6.0.2",
    "scikit-learn>=1.7.1",
]
